+++ packages/datawell/src/react/__specs__/index.spec.tsx
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react'
 import React from 'react'
 
-import DataWell from '..'
+import DataWell from '../index.js'
 
 describe('DataWell', () => {
   it('renders', () => {
